+++ packages/inventory/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @happyvertical/smrt-inventory
 
+## 0.29.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.29
+  - @happyvertical/smrt-tenancy@0.29.29
+
 ## 0.29.28
 
 ### Patch Changes