+++ packages/scanner/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @happyvertical/smrt-scanner
 
+## 0.29.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.29
+
 ## 0.29.28
 
 ### Patch Changes