+++ packages/subscriptions/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @happyvertical/smrt-subscriptions
 
+## 0.29.29
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.29.29
+  - @happyvertical/smrt-tenancy@0.29.29
+  - @happyvertical/smrt-svelte@0.29.29
+
 ## 0.29.28
 
 ### Patch Changes